Scene3DItem: fix assert when changing window
In Scene3DItem::updatePaintNode(), in the Render thread, we try to set
the parent of the aspect engine being deleted to nullptr. This was
triggering an assert in QCoreApplication::sendEvent(), since the aspect
engine lives in a different thread, the Main thread.

Don't trigger this assert by avoiding to send ChildAdded/ChildRemoved
events when reparenting the aspect manager, thanks to the
QQml_setParent_noEvent utility.

Pick-to: 6.4
Change-Id: Ic816701ee65654f7b18b4998c54feb4840af0a14
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>