RHIResourceManagers: fix the leak of RHIShader
The Renderer seems to already take care of deleting the abandoned
shaders thanks to the APIShaderManager::purge() method. However, when
Qt3D is shut down, there remain active RHIShaders in the
RHIShaderManager::m_apiShaders, that nobody deletes.

To fix this, in the RHIResourceManagers, delete the remaining RHIShaders
as we do with the other RHI resources.

Pick-to: 6.4
Change-Id: Iccc5470831eb7df0e120db6864fae425005c6919
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>