Doc: Replace \quotefromfile with \snippet to make code left-aligned
The \quotefromfile command is smart enough to keep the left indentation
from the source file when it renders the code in a code snippet in a
qdoc file. This is all good when you're listing multiple code snippets
and you can clearly see the left-most-aligned code, or 'starting
point'. However, when you use \quotefromfile for a single code snippet
that is indented in the source file, the rendered code in the qdoc file
is not left-aligned - and this looks a bit odd.

Pick-to: 6.5 6.4 6.2
Change-Id: Ie6d2445b37faa4a75749d6013001892de012a1ef
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>