Doc: Replace \quotefromfile with \snippet to make code left-aligned
The \quotefromfile command is smart enough to keep the left indentation
from the source file when it renders the code in a code snippet in a
qdoc file. This is all good when you're listing multiple code snippets
and you can clearly see the left-most-aligned code, or 'starting
point'. However, when you use \quotefromfile for a single code snippet
that is indented in the source file, the rendered code in the qdoc file
is not left-aligned - and this looks a bit odd.

Change-Id: Ie6d2445b37faa4a75749d6013001892de012a1ef
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 5859a840c3f21b36d4c7849c173c38e16ef52fd2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>