PlatformSurfaceFilter: fix crash when Scene3D is destroyed after window
When closing a window containing a Scene3D, the aspect engine is
destroyed after the QWindow (because of delete later), so the m_obj
pointer we store inside the PlatformSurfaceFilter becomes dangling,
leading to a crash in the destructor.

Use QPointer to ensure m_obj it set null when it is destroyed.

Fixes: QTBUG-111325
Change-Id: Icf153700a5708abff4b92b14ce2f9bb12d8653eb
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
(cherry picked from commit 2c3ebe01cb2875bfe48af20446b8c2b62ff49382)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>