Revert "Doc: Add tests/manual in exampledirs to fix broken links"
This reverts commit dee62f491a761cc860e5377746264d9e2990eca2.

Adding the manual tests as 'exampledirs' makes QDoc parse the example
documentation and also adds broken entries to the examples-manifest.xml.

We do not want the examples from /tests to generate documentation or
appear in Qt Creator. Revert the change, and address the QDoc warnings
that this commit tried to fix by other means.

Pick-to: 6.5
Change-Id: I848344ae0836e341fabda5b740162b0015f6baf1
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>