CMake: Make qt3d tests standalone projects
Add the boilerplate standalone test prelude to each test, so that they
can be opened with an IDE without the qt-cmake-standalone-test script,
but directly with qt-cmake or cmake.

Boilerplate was added using the following scripts:
https://git.qt.io/alcroito/cmake_refactor

Manual adjustments were made where the code was inserted in the wrong
location.

render/opengl tests were omitted because they include other files
that make the tests non-standalone.

Task-number: QTBUG-93020
Change-Id: I3c7e2ca8dafc15acfe909aad3fa60300ac9db525
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>