Fix issue with camera shaking with large offset values
If the case where the scene uses a QCamera, and that that
camera is not child of a node that has a transform component,
the renderview will use the viewMatrix computed from the
frontend. In all other cases, it will continue to recompute
the viewMatrix based on parameters of the camera and apply
parent transformations.

Task-number: QTBUG-69463
Pick-to: 6.7 6.6 6.5 5.15
Change-Id: I8300a4fd82c527ad3412e81b96633409794b3b1f
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>