RHI renderer: fix crash when executing a draw RV after a compute RV
RHI resource update batches are released when we pass them to
beginPass() or endPass(). We must then either clear the m_currentUpdates
pointer, or assign it to a new batch.

The later was done after all endPass(). But we forgot to clear the
m_currentUpdates pointer after beginPass(), which resulted in a crash,
since we then sent the same -released- batch pointer to the following
endPass().

This crash did not manifest before
6b6465032f103a084e416cbddf03f8e83191aea2 because then all batches were
leaked (never released).

Fixes: QTBUG-119657
Pick-to: 6.7
Change-Id: Ic63a96fc58714099b0a4e65e495ae0a720a3823f
Reviewed-by: Mike Krus <mike.krus@kdab.com>