Windows 7: blacklist tests that fail from low screen resolution
When OpenNebula was updated in CI, new Windows provisions defaulted to 800x600 resolution which caused some tests to fail. Adding new display settings to the Windows registry in tier-2 provisioning scripts worked on Windows 10/11 but issues with Windows 7 persists, one of my suspicions is that Windows 7 VM spawns a new empty video subkey in the registry when it's launched by a module in Coin run, which happens after provisioning scripts are run and done.

This issue is blocking Qt 5.15 integrations, so for now it is best to blacklist the tests that fail from the 800x600 screen resolution.

Task-number: QTQAINFRA-6145
Change-Id: Ib108c6de8ec6d5cd0b2601c50ce44cca2e432089
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Tarja Sundqvist <tarja.sundqvist@qt.io>