Rename QText2DMaterial uniform
Update the uniform name "textureSize" to "textureWidth" to avoid
name conflict with the glsl function textureSize.

This solves an issue where the shader transpiler mangles the
uniform name into "_textureSize" when using the RHI renderer with
the OpenGL backend. Causing the uniform to never be initialized
with the expected value and resulting in "jaggy" text.

Fixes: QTBUG-124708
Pick-to: 6.7
Change-Id: Ibe4b2aa0b86ee31b77594909843c6815f4da9fac
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>