quick3dscene3d: fix includes and LIBRARIES
Change includes to use the canonical form with <private/...> instead of
just <...>.

Also move some dependencies from LIBRARIES to PUBLIC_LIBRARIES, such
that users of the library don't get weird linking errors when linking to
that library. This is needed to be able to add a unit test to
quick3dscene3d, for example.
Remove the unneeded qglobal_p.h includes.

Task-number: QTBUG-129701
Change-Id: I0954078a371e41888b7252c8259afabe523c4c92
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>