CMake: Work around OpenGLRendererLib not having an SBOM type
The target is a bare library created with add_library(), and then uses
qt_internal_extend_target to add some attribution info.

This will break with a future qtbase that will try to generate an SBOM
for any target that adds any kind of sbom info, but doesn't provide an
SBOM TYPE.

qtbase SBOM support doesn't offer a good type for this kind of target
yet, so to work around it, we set the SBOM TYPE to CUSTOM for now.

Pick-to: 6.8 6.9 6.10
Task-number: QTBUG-134894
Change-Id: Ibd9101e6949c2f3d551927fa1523e8204a9cb249
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>